+++ crypto_sign/cross/clean/keccakf1600.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 void KeccakF1600_StateExtractBytes(uint64_t *state, unsigned char *data,
-                                   unsigned int offset, unsigned int length);
+                                   uint16_t offset, uint16_t length);
 void KeccakF1600_StateXORBytes(uint64_t *state, const unsigned char *data,
-                               unsigned int offset, unsigned int length);
+                               uint16_t offset, uint16_t length);
 void KeccakF1600_StatePermute(uint64_t *state);
